updates to upload, move mock eventsource server into middleware
