+++ interface/package.json
@@ -14,9 +14,7 @@
     "build-hosted": "typesafe-i18n --no-watch && vite build --mode hosted",
     "preview-standalone": "typesafe-i18n --no-watch && vite build && concurrently -c \"auto\" \"npm:mock-rest\" \"vite preview\"",
     "mock-rest": "bun --watch ../mock-api/rest_server.ts",
-    "mock-es": "bun --watch ../mock-api/es_server.ts",
-    "mock-upload": "bun --watch ../mock-api/upload_server.ts",
-    "standalone": "concurrently -c \"auto\" \"typesafe-i18n\" \"npm:mock-rest\" \"npm:mock-es\" \"npm:mock-upload\" \"vite\"",
+    "standalone": "concurrently -c \"auto\" \"typesafe-i18n\" \"npm:mock-rest\" \"vite\"",
     "typesafe-i18n": "typesafe-i18n --no-watch",
     "webUI": "node progmem-generator.js",
     "format": "prettier -l -w '**/*.{ts,tsx,js,css,json,md}'",
@@ -29,7 +27,7 @@
     "@mui/icons-material": "^5.16.7",
     "@mui/material": "^5.16.7",
     "@table-library/react-table-library": "4.1.7",
-    "alova": "3.0.8",
+    "alova": "3.0.9",
     "async-validator": "^4.2.5",
     "jwt-decode": "^4.0.0",
     "mime-types": "^2.1.35",
